# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CodeNexus - LeetCode-style coding platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /submissions:
    get:
      operationId: listSubmissions
      tags: [submissions]
      summary: List recent submissions for the current user
      parameters:
        - name: problemId
          in: query
          schema:
            type: integer
        - name: userId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Submission"
    post:
      operationId: createSubmission
      tags: [submissions]
      summary: Submit code for a problem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmissionInput"
      responses:
        "201":
          description: Submission created and judged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
  /submissions/{id}:
    get:
      operationId: getSubmission
      tags: [submissions]
      summary: Get a single submission detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submission detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
        "404":
          description: Not found
  /submissions/run:
    post:
      operationId: runCode
      tags: [submissions]
      summary: Run code against custom input (does not save)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunCodeInput"
      responses:
        "200":
          description: Run result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunResult"

  # ── Users ─────────────────────────────────────────────────
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users (leaderboard)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProfile"
    post:
      operationId: createUser
      tags: [users]
      summary: Create/register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
  /users/{id}/submissions:
    get:
      operationId: getUserSubmissions
      tags: [users]
      summary: Get all submissions for a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Submission"
  /users/{id}/stats:
    get:
      operationId: getUserStats
      tags: [users]
      summary: Get stats for a user (solved count, streak, acceptance)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserStats"
...
  /auth/login:
    post:
      operationId: login
      tags: [users]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
components:
  schemas:
...
    Submission:
      type: object
      required: [id, problemId, userId, language, code, verdict, runtime, memory, createdAt]
      properties:
        id:
          type: integer
        problemId:
          type: integer
        userId:
          type: integer
        language:
          type: string
        code:
          type: string
        verdict:
          type: string
          enum: [Accepted, Wrong Answer, Time Limit Exceeded, Memory Limit Exceeded, Compilation Error, Runtime Error]
        runtime:
          type: ["integer", "null"]
        memory:
          type: ["integer", "null"]
        testCasesPassed:
          type: ["integer", "null"]
        totalTestCases:
          type: ["integer", "null"]
        createdAt:
          type: string
        problemTitle:
          type: ["string", "null"]
        username:
          type: ["string", "null"]
    SubmissionInput:
      type: object
      required: [problemId, language, code]
      properties:
        problemId:
          type: integer
        language:
          type: string
        code:
          type: string
        userId:
          type: integer
    RunCodeInput:
      type: object
      required: [problemId, language, code]
      properties:
        problemId:
          type: integer
        language:
          type: string
        code:
          type: string
        customInput:
          type: string
    RunResult:
      type: object
      required: [output, status]
      properties:
        output:
          type: string
        status:
          type: string
        runtime:
          type: ["integer", "null"]
        error:
          type: ["string", "null"]

    # ── User schemas ──────────────────────────────────────────
    UserProfile:
      type: object
      required: [id, username, email, rating, rank, solvedCount, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        rating:
          type: integer
        rank:
          type: integer
        solvedCount:
          type: integer
        streak:
          type: integer
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        badges:
          type: array
          items:
            type: string
    UserInput:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 3
        email:
          type: string
        password:
          type: string
          minLength: 6
        bio:
          type: string
    UserStats:
      type: object
      required: [userId, solvedEasy, solvedMedium, solvedHard, totalSubmissions, acceptedSubmissions, streak, rating]
      properties:
        userId:
          type: integer
        solvedEasy:
          type: integer
        solvedMedium:
          type: integer
        solvedHard:
          type: integer
        totalSubmissions:
          type: integer
        acceptedSubmissions:
          type: integer
        streak:
          type: integer
        rating:
          type: integer
        languageStats:
          type: array
          items:
            $ref: "#/components/schemas/LanguageStat"
    LanguageStat:
      type: object
      required: [language, count]
      properties:
        language:
          type: string
        count:
          type: integer
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/UserProfile"
        token:
          type: string

    # ── Contest schemas ───────────────────────────────────────
...