# lib/api-zod/src/generated/types/languageStat.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CodeNexus - LeetCode-style coding platform API
 * OpenAPI spec version: 0.1.0
 */

export interface LanguageStat {
  language: string;
  count: number;
}